--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        rust: ["1.56.0", stable]
+        rust: ["1.60.0", stable]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v3