Visualized in `cargo build --timings`:

| Before: | After: |
|---|---|
| <img src="https://github.com/niklasmohrin/clircle/assets/1940490/0d74dd42-1afc-4320-9592-4460432e1ec7" width="350"> | <img src="https://github.com/niklasmohrin/clircle/assets/1940490/9af4aced-1ad1-41f5-8301-01ad8e985fa7" width="350"> |

Previously `serde` could only begin compiling after `serde_derive` was finished compiling. After, they can compile in parallel.

This reduces build time by 2 seconds (36%).